simaris now handles the events and target searching
